refactor: Move classes into separate files and add comments.
