+++ LibDmd/Output/Virtual/AlphaNumericResources.cs
@@ -154,7 +154,7 @@ namespace LibDmd.Output.Virtual
 				return null;
 			}
 			var surface = SKSurface.Create(dim.SvgInfo);
-			surface.Canvas.Translate(dim.TranslateX, dim.TranslateY);
+			surface.Canvas.Translate(dim.Translate.X, dim.Translate.Y);
 			Skew(surface.Canvas, skewAngle, 0);
 			foreach (var paint in paints) {
 				surface.Canvas.DrawPicture(segment, ref dim.SvgMatrix, paint);
@@ -251,207 +251,6 @@ namespace LibDmd.Output.Virtual
 	enum RasterizeLayer
 	{
 		OuterGlow, InnerGlow, Foreground, Background
-	}
-
-	public class RasterizeDimensions : IEquatable<RasterizeDimensions>
-	{
-		public int NumChars { get; set; }
-		public int NumLines { get; set; }
-
-		public int CanvasWidth { get; }
-		public int CanvasHeight { get; }
-
-		public float SvgWidth { get; }
-		public float SvgHeight { get; }
-		public float SvgScale { get; }
-		public SKImageInfo SvgInfo { get; }
-
-		public int OuterPadding { get; }
-		public int SegmentPadding { get; }
-		public int LinePadding { get; }
-
-		public float TranslateX => _svgSkewedWidth - SvgWidth + SegmentPadding;
-		public float TranslateY => SegmentPadding;
-
-		public SKMatrix SvgMatrix;
-
-		public float LinePaddingPercentage { get; set; } = 0.2f;
-		public float OuterPaddingPercentage { get; set; } = 0.2f;
-		public float SegmentPaddingPercentage { get; set; } = 0.3f;
-
-		private readonly float _svgSkewedWidth;
-
-		public RasterizeDimensions(SKRect svgSize, int canvasWidth, int canvasHeight, int numChars, int numLines, float skewAngle)
-		{
-			NumChars = numChars;
-			NumLines = numLines;
-
-			OuterPadding = (int)Math.Round(OuterPaddingPercentage * canvasHeight);
-			SvgHeight = canvasHeight - 2 * OuterPadding;
-			SvgScale = SvgHeight / svgSize.Height;
-			SvgWidth = svgSize.Width * SvgScale;
-			LinePadding = (int)Math.Round(SvgHeight * LinePaddingPercentage);
-			SvgMatrix = SKMatrix.MakeScale(SvgScale, SvgScale);
-			_svgSkewedWidth = SkewedWidth(SvgWidth, SvgHeight, skewAngle);
-			SegmentPadding = (int)Math.Round(Math.Sqrt(SvgWidth * SvgWidth + SvgHeight * SvgHeight) * SegmentPaddingPercentage);
-			SvgInfo = new SKImageInfo((int)(_svgSkewedWidth + 2 * SegmentPadding), (int)(SvgHeight + 2 * SegmentPadding));
-			var skewedWidth = SkewedWidth(SvgWidth, SvgHeight, skewAngle);
-			CanvasWidth = (int)Math.Round(2 * OuterPadding + (NumChars - 1) * SvgWidth + skewedWidth);
-			CanvasHeight = (int)Math.Round(OuterPadding * 2 + NumLines * SvgHeight + (NumLines - 1) * LinePadding);
-		}
-
-		private static float SkewedWidth(float width, float height, float angle)
-		{
-			var skew = (float)Math.Tan(Math.PI * angle / 180);
-			return width + Math.Abs(skew * height);
-		}
-
-		public bool Equals(RasterizeDimensions other)
-		{
-			if (ReferenceEquals(null, other)) return false;
-			if (ReferenceEquals(this, other)) return true;
-			return SvgHeight.Equals(other.SvgHeight);
-		}
-	}
-
-	public class DisplaySetting
-	{
-		public int Display { get; set; }
-		public SegmentType SegmentType { get; set; }
-		public RasterizeDimensions Dim { get; set; }
-		public RasterizeStyle Style {
-			get => _scaledStyle;
-			set { _style = value;
-				_scaledStyle = value.Scale(Dim);
-			}
-		}
-
-		public int NumLines { get; set; }
-		public int NumChars { get; set; }
-
-		private RasterizeStyle _style = new RasterizeStyle {
-			SkewAngle = -12,
-			Background = new RasterizeLayerStyle { Color = new SKColor(0xff, 0xff, 0xff, 0x20), Blur = new SKPoint(7, 7), IsEnabled = true, IsBlurEnabled = true, IsDilateEnabled = false },
-			OuterGlow = new RasterizeLayerStyle { Color = new SKColor(0xb6, 0x58, 0x29, 0x40), Blur = new SKPoint(50, 50), Dilate = new SKPoint(90, 40), IsEnabled = true, IsBlurEnabled = true, IsDilateEnabled = true },
-			InnerGlow = new RasterizeLayerStyle { Color = new SKColor(0xdd, 0x6a, 0x03, 0xa0), Blur = new SKPoint(15, 13), Dilate = new SKPoint(15, 10), IsEnabled = true, IsBlurEnabled = true, IsDilateEnabled = true },
-			Foreground = new RasterizeLayerStyle { Color = new SKColor(0xfb, 0xe6, 0xcb, 0xff), Blur = new SKPoint(2, 2), IsEnabled = true, IsBlurEnabled = true, IsDilateEnabled = false },
-		};
-		private RasterizeStyle _scaledStyle;
-
-		public void OnSvgsLoaded(SKRect svgSize, int canvasWidth, int canvasHeight)
-		{
-			Dim = new RasterizeDimensions(svgSize, canvasWidth, canvasHeight, NumChars, NumLines, _style.SkewAngle);
-			_scaledStyle = _style.Scale(Dim);
-		}
-	}
-
-	public class RasterizeStyle
-	{
-		public float SkewAngle { get; set; }
-		public RasterizeLayerStyle Foreground { get; set; }
-		public RasterizeLayerStyle InnerGlow { get; set; }
-		public RasterizeLayerStyle OuterGlow { get; set; }
-		public RasterizeLayerStyle Background { get; set; }
-
-		public RasterizeStyle Scale(RasterizeDimensions dim)
-		{
-			return new RasterizeStyle {
-				SkewAngle = SkewAngle,
-				Foreground = Foreground.Scale(dim),
-				InnerGlow = InnerGlow.Scale(dim),
-				OuterGlow = OuterGlow.Scale(dim),
-				Background = Background.Scale(dim),
-			};
-		}
-
-		public RasterizeStyle Copy()
-		{
-			return new RasterizeStyle {
-				SkewAngle = SkewAngle,
-				Foreground = Foreground.Copy(),
-				InnerGlow = InnerGlow.Copy(),
-				OuterGlow = OuterGlow.Copy(),
-				Background = Background.Copy()
-			};
-		}
-	}
-
-	public class RasterizeLayerStyle
-	{
-		public bool IsEnabled { get; set; }
-		public bool IsBlurEnabled { get; set; }
-		public bool IsDilateEnabled { get; set; }
-		public SKColor Color { get; set; }
-		public SKPoint Blur { get; set; }
-		public SKPoint Dilate { get; set; }
-
-		public RasterizeLayerStyle Scale(RasterizeDimensions dim)
-		{
-			return new RasterizeLayerStyle {
-				IsEnabled = IsEnabled,
-				IsBlurEnabled = IsBlurEnabled,
-				IsDilateEnabled = IsDilateEnabled,
-				Color = Color,
-				Blur = new SKPoint(dim.SvgScale * Blur.X, dim.SvgScale * Blur.Y),
-				Dilate = new SKPoint((float) Math.Round(dim.SvgScale * Dilate.X), (float) Math.Round(dim.SvgScale * Dilate.Y))
-			};
-		}
-
-		public RasterizeLayerStyle Copy()
-		{
-			return new RasterizeLayerStyle {
-				IsEnabled = IsEnabled,
-				IsBlurEnabled = IsBlurEnabled,
-				IsDilateEnabled = IsDilateEnabled,
-				Color = new SKColor(Color.Red, Color.Green, Color.Blue, Color.Alpha),
-				Blur = new SKPoint(Blur.X, Blur.Y),
-				Dilate = new SKPoint(Dilate.X, Dilate.Y)
-			};
-		}
-
-		public override bool Equals(object obj)
-		{
-			if (!(obj is RasterizeLayerStyle item)) {
-				return false;
-			}
-
-			return IsEnabled == item.IsEnabled
-			       && IsBlurEnabled == item.IsBlurEnabled
-			       && IsDilateEnabled == item.IsDilateEnabled
-			       && Color.Equals(item.Color)
-			       && Blur.Equals(item.Blur)
-			       && Dilate.Equals(item.Dilate);
-		}
-
-		protected bool Equals(RasterizeLayerStyle other)
-		{
-			return IsEnabled == other.IsEnabled 
-			       && IsBlurEnabled == other.IsBlurEnabled 
-			       && IsDilateEnabled == other.IsDilateEnabled 
-			       && Color.Equals(other.Color) 
-			       && Blur.Equals(other.Blur) 
-			       && Dilate.Equals(other.Dilate);
-		}
-
-		public override int GetHashCode()
-		{
-			unchecked
-			{
-				var hashCode = IsEnabled.GetHashCode();
-				hashCode = (hashCode * 397) ^ IsBlurEnabled.GetHashCode();
-				hashCode = (hashCode * 397) ^ IsDilateEnabled.GetHashCode();
-				hashCode = (hashCode * 397) ^ Color.GetHashCode();
-				hashCode = (hashCode * 397) ^ Blur.GetHashCode();
-				hashCode = (hashCode * 397) ^ Dilate.GetHashCode();
-				return hashCode;
-			}
-		}
-
-		public override string ToString()
-		{
-			return
-				$"LayerStyle[enabled:{IsEnabled},color:{Color.ToString()},blur:{IsBlurEnabled}/{Blur.X}x{Blur.Y},dilate:{IsDilateEnabled}/{Dilate.X}x{Dilate.Y}";
-		}
 	}
 
 	struct RasterCacheKey
